Fix issues with some clients not accepting the agent/eliminated the timing issue
It turns out that if you read too much data from the pipe, OpenSSH observes this (intentionally or not, I'm not sure). This was done by an internal buffer when calling `ReadAsync()`. Switching to `ReadExactlyAsync()` prevents the overread.

Also moved client processing code into its own function, and wrapped it in a try/catch/finally to correctly handle end of streams and cancel the token for any additional operations, such as a UI prompt